+++ yarn-project/aztec.js/src/account_impl/account_collection.ts
@@ -6,7 +6,7 @@ import { AccountImplementation } from './index.js';
  * A concrete account implementation that manages multiple accounts.
  */
 export class AccountCollection implements AccountImplementation {
-  private accounts: Map<string, AccountImplementation> = new Map();
+  private accounts: Map<AztecAddress, AccountImplementation> = new Map();
 
   /**
    * Registers an account implementation against an aztec address
@@ -14,12 +14,12 @@ export class AccountCollection implements AccountImplementation {
    * @param impl - The account implementation to be registered.
    */
   public registerAccount(addr: AztecAddress, impl: AccountImplementation) {
-    this.accounts.set(addr.toString(), impl);
+    this.accounts.set(addr, impl);
   }
 
   getAddress(): AztecAddress {
     if (!this.accounts) throw new Error(`No accounts registered`);
-    return AztecAddress.fromString(this.accounts.keys().next().value as string);
+    return this.accounts.keys().next().value as AztecAddress;
   }
 
   /**
@@ -34,7 +34,7 @@ export class AccountCollection implements AccountImplementation {
   ): Promise<TxExecutionRequest> {
     // TODO: Check all executions have the same origin
     const sender = executions[0].from;
-    const impl = this.accounts.get(sender.toString());
+    const impl = this.accounts.get(sender);
     if (!impl) throw new Error(`No account implementation registered for ${sender}`);
     return impl.createAuthenticatedTxRequest(executions, txContext);
   }
